Refactor (APIMF-3491): big step -> extracted node parsing from DialectInstanceParser
