+++ amf-aml/shared/src/main/scala/amf/aml/internal/parse/instances/DialectInstanceFragmentParser.scala
@@ -7,6 +7,7 @@ import amf.aml.internal.metamodel.document.DialectInstanceModel
 import amf.aml.client.scala.model.document.{DialectInstanceFragment, DialectInstanceProcessingData}
 import amf.aml.client.scala.model.domain.{DialectDomainElement, DocumentsModel}
 import amf.aml.internal.parse.instances.DialectInstanceParser.encodedElementDefaultId
+import amf.aml.internal.parse.instances.parser.InstanceNodeParser
 import amf.aml.internal.validate.DialectValidations.DialectError
 import com.github.ghik.silencer.silent
 
@@ -46,7 +47,7 @@ class DialectInstanceFragmentParser(root: Root)(implicit override val ctx: Diale
         ctx.findNodeMapping(documentMapping.encoded().value()) match {
           case Some(nodeMapping) =>
             val path = dialectInstanceFragment.id + "#"
-            parseNode(path, path + "/", map, nodeMapping, Map(), givenAnnotations = None)
+            InstanceNodeParser(root).parse(path, path + "/", map, nodeMapping, Map(), givenAnnotations = None)
           case _ =>
             emptyElementWithViolation(s"Could not find node mapping for: ${documentMapping.encoded().value()}")
         }
